[Editorial] Replace publisher's "[sic]" text with corrected text
The 1905 Funk & Wagnalls Co. edition preserved several mistakes that
were in the first edition, but called out their presence with
italicized "[sic]" text.

These mistakes were corrected in subsequent print runs after the first
edition, so this commit replaces the "[sic]" text with the fixed text
that became part of other editions.

The page scans from two previous editions were consulted:
  - H. Altemus Company (1900)
  - W. A. Leary & Co (1883)

Another possible typo is called out in the author's preface, but it
wasn't corrected in the above editions (or any other editions that I
could find) so the "[sic]" designation has been removed from that file
for the sake of consistency.